+++ packages/angular-components/CHANGELOG.md
@@ -7,4 +7,4 @@
 - [#133](https://github.com/swisspost/design-system/pull/133) [`b380ded`](https://github.com/swisspost/design-system/commit/b380dedefc6a88626b2ee0706efefe438b519d3b) Thanks [@gfellerph](https://github.com/gfellerph)! - Endeavour the 1st, systems are go
 
 - Updated dependencies [[`b380ded`](https://github.com/swisspost/design-system/commit/b380dedefc6a88626b2ee0706efefe438b519d3b)]:
-  - @swisspost/web-styles@5.0.0-alpha.2
+  - @swisspost/design-system-styles@5.0.0-alpha.2